updating the event
