--- packages/spruce-skill/server/events/get-cards.js
@@ -1,22 +1,0 @@
-const config = require('config')
-const { eventError } = require('../lib/errorHandler')
-
-module.exports = async (ctx, next) => {
-	try {
-		log.debug('** Event: get-page-cards **', { event: ctx.event })
-
-		// Determine which cards should be sent to the user
-		const cards = []
-
-		cards.push(config.cards.exampleCard)
-
-		ctx.body = cards
-		await next()
-	} catch (e) {
-		eventError({
-			ctx,
-			next,
-			e
-		})
-	}
-}
